+++ R/waldo-package.R
@@ -8,3 +8,5 @@
 ## usethis namespace: start
 ## usethis namespace: end
 NULL
+
+release_extra_revdeps <- function() "testthat"