+++ tests/testthat/test-compare-value.R
@@ -112,6 +112,14 @@ test_that("shows row-by-row diff for numeric matrices", {
   })
 })
 
+test_that("but not for arrays", {
+  expect_snapshot({
+    x <- y <- array(1:4, c(1, 2, 2))
+    y[1, 2, 2] <- 5L
+    compare(x, y)
+  })
+})
+
 test_that("falls back to regular display if printed representation the same", {
   expect_snapshot({
     x <- y <- matrix(1:4, nrow = 2)
@@ -141,6 +149,13 @@ test_that("don't use format if numeric & within tolerance", {
   expect_snapshot({
     compare(dt, dt + 5)
     compare(dt, dt + 5, tolerance = 1e-8)
+  })
+})
+
+test_that("can compare complex numbers", {
+  expect_snapshot({
+    compare(1:2 + 1i, 2 + 1i)
+    compare(1:2 + 1i, 1:2 + 2i)
   })
 })