+++ R/compare-opts.R
@@ -6,7 +6,8 @@ compare_opts <- function(...,
                          ignore_encoding = TRUE,
                          ignore_function_env = FALSE,
                          ignore_formula_env = FALSE,
-                         list_as_map = FALSE
+                         list_as_map = FALSE,
+                         quote_strings = TRUE
                          ) {
 
   base <- old_opts(...)
@@ -23,6 +24,7 @@ compare_opts <- function(...,
     ignore_function_env = ignore_function_env,
     ignore_formula_env = ignore_formula_env,
     list_as_map = list_as_map,
+    quote_strings = quote_strings,
     seen = seen
   )