+++ NEWS.md
@@ -1,3 +1,5 @@
+# waldo (development version)
+
 # waldo 0.4.0
 
 * Atomic S3 classes with format methods now use those methods when