+++ NEWS.md
@@ -3,6 +3,12 @@
 * `compare()` ignores the `balanced` attribute used by some `POSIXlt` objects
   in R 4.3 and greater (#160).
 
+* `compare()` once again works with 3d (and greater) numeric arrays (#148).
+
+* `compare()` gains basic support for complex numbers (#146).
+
+* `ignore_attr = "class"` now works for more types of input (#143).
+
 * `compare()` now correctly reports that `NA_real_` and `NaN` are different
   (@sorhawell, #150).