+++ NEWS.md
@@ -3,6 +3,9 @@
 * Leading and trailing `NA`s are no longer omitted from output when the
   lengths of `x` and `y` are unequal (#109).
 
+* `compare()` can now opt-out of quoting strings with `quote_strings = FALSE` 
+  (#145).
+
 * `NA_character_` and `"NA"` no longer compare as equal (#162).
 
 * `compare()` ignores the `balanced` attribute used by some `POSIXlt` objects